Adds bootstrap plotting to MCMC comparison
Adds functionality to plot bootstrap estimates alongside flexible and constant MCMC results.
This includes plotting the bootstrap median as a horizontal line and the confidence interval as a shaded area.
Updates the plot total function to incorporate the bootstrap data and modifies the pipeline to accept and pass bootstrap values.
Also, renames the action colorbar label to step size.